requirements.txt
djangorestframework==3.15.2
Django==5.0.6
drf-yasg==1.21.7
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
django-cors-headers==4.3.1